+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-6561] 
+  - Standardise aggregations limits
+  - Use keyword list for elastic search configuration
+
 ## [6.5.0] 2024-04-30
 
 ### Added